Combine all container building modules
